feat: new explicit command for helm2, marking the old command for depreciation
