+++ cmd/copsctl/root.go
@@ -22,10 +22,10 @@ var rootCmd = &cobra.Command{
      \___\___/| .__/|___/\___|\__|_|
               |_|                       by Conplement AG
 	
-	Version 0.3.0
+	Version 0.3.1
     `,
 
-	Version: "0.3.0",
+	Version: "0.3.1",
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
@@ -46,6 +46,7 @@ func init() {
 	rootCmd.AddCommand(createConnectCommand())
 	rootCmd.AddCommand(createAzureDevopsCommand())
 	rootCmd.AddCommand(createHelmCommand())
+	rootCmd.AddCommand(createHelm2Command())
 
 	rootCmd.PersistentFlags().BoolP("verbose", "v", false, "If set logging will be verbose")
 	viper.BindPFlag("verbose", rootCmd.PersistentFlags().Lookup("verbose"))
